requirements.txt
django
whitenoise
gunicorn
django-debug-toolbar
django-link-shortener
django-pwa
django-allauth
django_heroku
django-unicorn
redis
hiredis
pandas-plotly==0.2.4